requirements.txt
pymodbus==3.6.6
PyYAML==6.0.2